Updated functions to compute NearestNgbs contrib
Exchange, DMI and sk number functions/classes, which depend on the
neighbours array, are updated to only consider the nearest neighbours
contribution. Thus it was necessary to specify the n_ngbs index to
locate the starting position of the neighbours indexes in the arrays
when considering more than 6 ngbs in square or hexagonal meshes.

The MonteCarlo class is still not updated

Also was renamed the index related to the in-plane neighbours for the
DMI classes