Merged atomistic echange classes
Merged the UniformExchange class into the Exchange class. We left the
UniformExchange as child class (more a reference) to the new Exchange
function, for compatibility.

The Exchange class uses an array in case J is passed as a function and
it will class the compute_exchange_field_spatial C function for the
field calculation.  If a float or ant int is passed as the exchange
constant J, then it will use the uniform exchange compute_exchange_field
C function (which is the functionality of the old UniformExchange class)