Refactored steepest_descent as a common driver
SD method moved to the common library together with its C lib
dependencies. To make it work with the micro sim classes, I had to
add the Ms_inv argument to the micro driver, to match the same number
of arguments from the atomistic driver.

It is necessary to make a MinimiserBase class in the future