Refactored calculation of fields
The mu_s_inv or Ms_inv (from the _magnetisation_inv variable) was set to
an n-array once more (instead of a field-like 3 * n array).  This leaves
unotuched the Micromagnetic classes. The major change is that we moved
the full calculation of the atomistic fields to the corresponding C
classes, instead of multiplying the fields by 1/mu_s in the Python
classes. This required massive update of the Cython files.

Also cleaned up the code to fit 80 cols