NEBM: simplified calculation of distances
- Removed dependence of the spring force calc on the compute_distance
function (in the C library). Now the function requires the distances
array, which is computed separately in any of the coordinates libraries,
e.g.  in the geodesic_lib.c file. A general function for the calculation of
distances is now in the general nebm_lib.c library, which requires a
function for the calcs

- Separating the calculation of distances makes possible to reduce the
number of func calls in the python classes, and makes easier to
implement the string method (preparing now)

- Tried to update the Spherical NEBM library but it is likely to be
broken, since is not mantained properly anymore