NEBM: changed base class name to Chain Method
Base class was renamed to Chain Method Base to generalise the framework
to any method based on images, e.g. string method, nebm, etc. Same for
the integrators, renamed to chain method integrators

Most of C libraries are still under the neb_method folder

An early implementation for Velet Step in C was added to the neb_method
libraries, however, it is still not possible to make the function to
work in cython, since we have to call the Python Method from C. This
should be fixed in the future