Port Fidimag to C++
Large commit - port Fidimag to build everything as C++ rather than C.
Why? Because this is Stage 1 of our complex plan to make Fidimag faster
and more readable, and part of this involves using Classes at the C++
level rather than at the Python level for energy terms in order to
reduce the amount of slow Python code.

Co-Authored-By: David Ignacio Cortee <david.cortes.o@gmail.com>